# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Q-Vuln Post-Quantum Security Scanner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scans:
    post:
      operationId: createScan
      tags: [scans]
      summary: Start a new scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanBody"
      responses:
        "201":
          description: Scan initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{id}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "404":
          description: Not found
  /scans/{id}/findings:
    get:
      operationId: getScanFindings
      tags: [scans]
      summary: Get all findings for a scan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan findings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
...
components:
  schemas:
...
    Scan:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        mode:
          type: string
          enum: [scan-only, interactive, proactive, community]
        status:
          type: string
          enum: [pending, running, completed, failed]
        riskScore:
          type: integer
          nullable: true
        totalLines:
          type: integer
        criticalCount:
          type: integer
        alertCount:
          type: integer
        cleanCount:
          type: integer
        totalEffortHours:
          type: number
        estimatedCost:
          type: integer
        executiveSummary:
          type: string
          nullable: true
        findings:
          type: array
          items:
            $ref: "#/components/schemas/Finding"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - projectId
        - mode
        - status
        - totalLines
        - criticalCount
        - alertCount
        - cleanCount
        - totalEffortHours
        - estimatedCost
        - createdAt
    CreateScanBody:
      type: object
      properties:
        projectId:
          type: integer
        mode:
          type: string
          enum: [scan-only, interactive, proactive, community]
        code:
          type: string
        language:
          type: string
      required:
        - projectId
        - mode
        - code
        - language
    Finding:
      type: object
      properties:
        id:
          type: integer
        scanId:
          type: integer
        fileName:
          type: string
        lineNumber:
          type: integer
        severity:
          type: string
          enum: [critical, alert, safe]
        algorithm:
          type: string
        codeSnippet:
          type: string
        nistReplacement:
          type: string
          nullable: true
        nistStandard:
          type: string
          nullable: true
        effortHours:
          type: number
        explanation:
          type: string
          nullable: true
      required:
        - id
        - scanId
        - fileName
        - lineNumber
        - severity
        - algorithm
        - codeSnippet
        - effortHours
...